# refactor(list_archives): Reduce calls to date.format()