--- a/lib/plugins/helper/list_archives.js
+++ b/lib/plugins/helper/list_archives.js
@@ -13,6 +13,10 @@ function listArchivesHelper(options = {}) {
   const showCount = Object.prototype.hasOwnProperty.call(options, 'show_count') ? options.show_count : true;
   const className = options.class || 'archive';
   const order = options.order || -1;
+  const compareFunc = type === 'monthly'
+    ? (yearA, monthA, yearB, monthB) => yearA === yearB && monthA === monthB
+    : (yearA, monthA, yearB, monthB) => yearA === yearB;
+
   let result = '';
 
   if (!format) {
@@ -30,14 +34,14 @@ function listArchivesHelper(options = {}) {
     let date = post.date.clone();
 
     if (timezone) date = date.tz(timezone);
-    if (lang) date = date.locale(lang);
 
     const year = date.year();
     const month = date.month() + 1;
-    const name = date.format(format);
     const lastData = data[length - 1];
 
-    if (!lastData || lastData.name !== name) {
+    if (!lastData || !compareFunc(lastData.year, lastData.month, year, month)) {
+      if (lang) date = date.locale(lang);
+      const name = date.format(format);
       length = data.push({
         name,
         year,