--- a/test/scripts/helpers/list_archives.js
+++ b/test/scripts/helpers/list_archives.js
@@ -93,6 +93,19 @@ describe('list_archives', () => {
     ].join(''));
   });
 
+  it('show_count + style: false', () => {
+    const result = listArchives({
+      style: false,
+      show_count: false
+    });
+
+    result.should.eql([
+      '<a class="archive-link" href="/archives/2014/02/">February 2014</a>',
+      '<a class="archive-link" href="/archives/2013/10/">October 2013</a>',
+      '<a class="archive-link" href="/archives/2013/06/">June 2013</a>'
+    ].join(', '));
+  });
+
   it('order', () => {
     const result = listArchives({
       order: 1
@@ -121,6 +134,21 @@ describe('list_archives', () => {
       '<li class="archive-list-item"><a class="archive-list-link" href="/archives/2013/06/">JUNE 2013</a><span class="archive-list-count">2</span></li>',
       '</ul>'
     ].join(''));
+  });
+
+  it('transform + style: false', () => {
+    const result = listArchives({
+      style: false,
+      transform(str) {
+        return str.toUpperCase();
+      }
+    });
+
+    result.should.eql([
+      '<a class="archive-link" href="/archives/2014/02/">FEBRUARY 2014<span class="archive-count">1</span></a>',
+      '<a class="archive-link" href="/archives/2013/10/">OCTOBER 2013<span class="archive-count">1</span></a>',
+      '<a class="archive-link" href="/archives/2013/06/">JUNE 2013<span class="archive-count">2</span></a>'
+    ].join(', '));
   });
 
   it('separator', () => {