<!--
Thank you for creating a pull request to contribute to Hexo code! Before you open the request please answer the following questions to help it be more easily integrated. Please check the boxes "[ ]" with "[x]" when done too.
-->

## What does it do?

Continuation of #4007.

`list_archives` will call `date.locale(lang)`, `date.format(format)` a lot of times, especial if set to `relative_link: true`. (fragment_cache was disabled)

This PR change makes `list_archives` call `date.locale(lang)`, `date.format(format)` once per group.

## How to test

```sh
git clone -b improve-list_archives https://github.com/dailyrandomphoto/hexo.git
cd hexo
npm install
npm test
```


## Pull request tasks

- [x] Add test cases for the changes.
- [x] Passed the CI test.