@dailyrandomphoto 2019-12-26
A benchmark using a hexo dummy website (with 900 posts):

Measure 5 times and calculate the average for each benchmark.

Benchmark conditions:
- common:
  - language: zh-CN
  - **relative_link: true**
- master-1:
  - version: Before #4007. [commit](https://github.com/dailyrandomphoto/hexo/commit/92c705dd33f1545ba3eb5dc7c1bad98ebc3ae667)
- master-2:
  - version: Apply #4007. [commit](https://github.com/dailyrandomphoto/hexo/commit/75fea56bd88e5d3702c0d3813836cd8392849dc2)


**Node.js 8**

| | Total time (Cold) | Save DB (Cold) | Memory (Cold) | Total time (Hot) | Save DB (Hot) | Memory (Hot) |
|:---:|:---:|:---:|:---:|:---:|:---:|:---:|
| master-1 | 148.839s | 2.38s | 1102.397MB | 125.162s | 2.111s | 1181.39MB |
| master-2 | 81.695s | 2.219s | 1112.943MB | 63.956s | 1.973s | 1172.119MB |
| this PR | 77.795s | 2.152s | 1109.003MB | 59.523s | 1.956s | 1174.696MB |


**Node.js 10**

| | Total time (Cold) | Save DB (Cold) | Memory (Cold) | Total time (Hot) | Save DB (Hot) | Memory (Hot) |
|:---:|:---:|:---:|:---:|:---:|:---:|:---:|
| master-1 | 121.885s | 2.045s | 1203.274MB | 106.81s | 1.859s | 1242.954MB |
| master-2 | 68.21s | 2.045s | 1167.641MB | 53.912s | 1.888s | 1206.12MB |
| this PR | 67.258s | 2.039s | 1189.209MB | 53.68s | 1.896s | 1203.623MB |


**Node.js 12**

| | Total time (Cold) | Save DB (Cold) | Memory (Cold) | Total time (Hot) | Save DB (Hot) | Memory (Hot) |
|:---:|:---:|:---:|:---:|:---:|:---:|:---:|
| master-1 | 127.652s | 2.065s | 1128.723MB | 114.08s | 1.953s | 1210.77MB |
| master-2 | 69.073s | 2.112s | 1141.249MB | 54.851s | 1.945s | 1183.367MB |
| this PR | 66.376s | 2.084s | 1146.682MB | 51.547s | 1.994s | 1169.289MB |


**Node.js 13**

| | Total time (Cold) | Save DB (Cold) | Memory (Cold) | Total time (Hot) | Save DB (Hot) | Memory (Hot) |
|:---:|:---:|:---:|:---:|:---:|:---:|:---:|
| master-1 | 125.619s | 2s | 1129.541MB | 114.066s | 1.898s | 1181.748MB |
| master-2 | 69.611s | 2.1s | 1125.007MB | 54.739s | 1.915s | 1158.469MB |
| this PR | 68.913s | 2.034s | 1116.61MB | 53.455s | 1.913s | 1168.279MB |